requirements.txt
<<<<<<< HEAD
telegram
configparser
redis
python-dotenv==0.19.2
=======
anyio==4.2.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
configparser==6.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
python-telegram-bot==13.7
pytz==2024.1
redis
requests==2.31.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
telegram==0.0.1
tornado==6.4
tzlocal==5.2
urllib3==1.26.18
wheel==0.41.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
sqlalchemy
>>>>>>> origin/simon-push
